% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{.coords_to_svg_path}
\alias{.coords_to_svg_path}
\title{Convert coordinates to SVG path string}
\usage{
.coords_to_svg_path(coords)
}
\description{
Convert coordinates to SVG path string
}
\keyword{internal}
