% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery.R
\name{.download_tiles}
\alias{.download_tiles}
\title{Download and merge tiles into GeoTIFF}
\usage{
.download_tiles(bbox, output, source, zoom, api_key)
}
\description{
Download and merge tiles into GeoTIFF
}
\keyword{internal}
