% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{.extract_polygon_coords}
\alias{.extract_polygon_coords}
\title{Extract polygon coordinates for plotting}
\usage{
.extract_polygon_coords(geom, img_height)
}
\description{
Extract polygon coordinates for plotting
}
\keyword{internal}
