% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.fake_geo_to_pixel}
\alias{.fake_geo_to_pixel}
\title{Convert Fake Geographic to Pixel Coordinates}
\usage{
.fake_geo_to_pixel(lng, lat, height)
}
\arguments{
\item{lng}{Longitude (fake)}

\item{lat}{Latitude (fake)}

\item{height}{Image height in pixels}
}
\value{
Named vector with x, y (pixel coords, y from top)
}
\description{
Convert Fake Geographic to Pixel Coordinates
}
\keyword{internal}
