% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{.geo_boxes_to_pixel}
\alias{.geo_boxes_to_pixel}
\title{Convert Geographic Boxes to Pixel Boxes}
\usage{
.geo_boxes_to_pixel(boxes, image_path, input_crs = 4326)
}
\arguments{
\item{boxes}{List of bounding boxes (xmin, ymin, xmax, ymax in geographic CRS)}

\item{image_path}{Path to the reference GeoTIFF}

\item{input_crs}{CRS of input coordinates (default WGS84)}
}
\value{
List of pixel boxes (xmin, ymin, xmax, ymax in pixels)
}
\description{
Internal function to convert geographic bounding boxes to pixel coordinates.
}
\keyword{internal}
