% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.get_mapbox_token}
\alias{.get_mapbox_token}
\title{Get Mapbox Token}
\usage{
.get_mapbox_token()
}
\value{
Character string with the token, or NULL if not found.
}
\description{
Internal function to get the Mapbox access token from various sources.
}
\keyword{internal}
