% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{.get_module}
\alias{.get_module}
\title{Get Python Module}
\usage{
.get_module()
}
\value{
The Python geosam_core module.
}
\description{
Internal function to get the initialized Python module.
}
\keyword{internal}
