% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{.get_pytorch_packages}
\alias{.get_pytorch_packages}
\title{Get PyTorch packages based on GPU availability}
\usage{
.get_pytorch_packages(gpu)
}
\description{
Get PyTorch packages based on GPU availability
}
\keyword{internal}
