% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.image_extent}
\alias{.image_extent}
\title{Get Fake Geographic Extent for Image}
\usage{
.image_extent(width, height)
}
\arguments{
\item{width}{Image width in pixels}

\item{height}{Image height in pixels}
}
\value{
Named list with xmin, xmax, ymin, ymax
}
\description{
Returns the fake WGS84 extent used for displaying an image in MapLibre.
}
\keyword{internal}
