% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{.install_virtualenv}
\alias{.install_virtualenv}
\title{Install using virtualenv (via reticulate)}
\usage{
.install_virtualenv(envname, gpu, python_version)
}
\description{
Install using virtualenv (via reticulate)
}
\keyword{internal}
