% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.normalize_boxes}
\alias{.normalize_boxes}
\title{Normalize Boxes Input}
\usage{
.normalize_boxes(boxes)
}
\arguments{
\item{boxes}{Matrix, data.frame, or vector of box coordinates}
}
\value{
List of \link{xmin, ymin, xmax, ymax} vectors for Python
}
\description{
Convert various box formats to a consistent matrix format.
}
\keyword{internal}
