% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery.R
\name{.notify_tos}
\alias{.notify_tos}
\title{Notify user of Terms of Service (once per session per provider)}
\usage{
.notify_tos(source)
}
\description{
Notify user of Terms of Service (once per session per provider)
}
\keyword{internal}
