% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.pixel_to_fake_geo}
\alias{.pixel_to_fake_geo}
\title{Convert Pixel Coordinates to Fake Geographic}
\usage{
.pixel_to_fake_geo(x, y, height)
}
\arguments{
\item{x}{X pixel coordinate (from left)}

\item{y}{Y pixel coordinate (from top)}

\item{height}{Image height in pixels}
}
\value{
Named vector with lng, lat
}
\description{
Convert Pixel Coordinates to Fake Geographic
}
\keyword{internal}
