% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{.read_geojson}
\alias{.read_geojson}
\title{Read GeoJSON Result}
\usage{
.read_geojson(path)
}
\arguments{
\item{path}{Path to GeoJSON file}
}
\value{
sf object or NULL if empty/error
}
\description{
Internal function to read detection results from GeoJSON.
}
\keyword{internal}
