% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{.read_image_array_magick}
\alias{.read_image_array_magick}
\title{Read Image as Array using magick}
\usage{
.read_image_array_magick(img)
}
\arguments{
\item{img}{A magick image object}
}
\value{
Numeric array (height x width x 3)
}
\description{
Internal function to read any image format and return as RGB array for Python.
}
\keyword{internal}
