% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{.sam_as_sf_image}
\alias{.sam_as_sf_image}
\title{Convert geosam_image masks to sf (pixel coordinates)}
\usage{
.sam_as_sf_image(x, min_area = NULL, max_area = NULL)
}
\description{
Convert geosam_image masks to sf (pixel coordinates)
}
\keyword{internal}
