% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery.R
\name{.tile_to_lonlat}
\alias{.tile_to_lonlat}
\title{Convert tile coordinates to lon/lat (NW corner)}
\usage{
.tile_to_lonlat(x, y, zoom)
}
\description{
Convert tile coordinates to lon/lat (NW corner)
}
\keyword{internal}
