% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_geosam}
\alias{is_geosam}
\title{Check if Object is a geosam}
\usage{
is_geosam(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} is a geosam object.
}
\description{
Check if Object is a geosam
}
