% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_geosam_image}
\alias{is_geosam_image}
\title{Check if Object is a geosam_image}
\usage{
is_geosam_image(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} is a geosam_image object.
}
\description{
Check if Object is a geosam_image
}
