% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.geosam}
\alias{print.geosam}
\title{Print Method for geosam Objects}
\usage{
\method{print}{geosam}(x, ...)
}
\arguments{
\item{x}{A geosam object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for geosam Objects
}
