% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.geosam_image}
\alias{print.geosam_image}
\title{Print Method for geosam_image Objects}
\usage{
\method{print}{geosam_image}(x, ...)
}
\arguments{
\item{x}{A geosam_image object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for geosam_image Objects
}
