% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{sam_count}
\alias{sam_count}
\title{Get Number of Detections}
\usage{
sam_count(x)
}
\arguments{
\item{x}{A geosam object.}
}
\value{
Integer count of detections.
}
\description{
Returns the number of detected objects.
}
