% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{sam_scores}
\alias{sam_scores}
\title{Get Detection Scores}
\usage{
sam_scores(x)
}
\arguments{
\item{x}{A geosam or geosam_image object.}
}
\value{
Numeric vector of scores.
}
\description{
Returns confidence scores for all detections.
}
